separate profile and email controllers
